refactor: (피드백 반영) 다운 캐스팅으로 인한 오버플로우 방지
- 로또 구매 상한선 상수 및 예외 처리 추가
- 금액 관련 변수를 long으로 수정

+++ src/main/java/lotto/domain/Lottos.java
@@ -27,9 +27,9 @@ public class Lottos {
 
     // 로또 리스트들의 당첨금 총액 반환
     public Money getLottoSum() {
-        int sum = 0;
+        long sum = 0;
         for (Lotto lotto : lottoList) {
-            sum += (int) lotto.getLottoRank().getValue();
+            sum += lotto.getLottoRank().getValue();
         }
 
         return Money.from(sum);
